requirements.txt
django==4.1
Fabric==1.10.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
